Moving snackbar logic to effects
