+++ libs/books/feature/src/lib/book-search/book-search.component.ts
@@ -5,14 +5,9 @@ import {
   clearSearch,
   getAllBooks,
   searchBooks,
-  removeFromReadingList,
-  confirmedAddToReadingList
 } from '@tmo/books/data-access';
 import { FormBuilder } from '@angular/forms';
 import { Book } from '@tmo/shared/models';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { BookConstant } from '../book.constant';
-import { take } from 'rxjs/operators';
 
 @Component({
   selector: 'tmo-book-search',
@@ -28,8 +23,7 @@ export class BookSearchComponent {
 
   constructor(
     private readonly store: Store,
-    private readonly fb: FormBuilder,
-    private snackbar: MatSnackBar
+    private readonly fb: FormBuilder
   ) {}
 
   get searchTerm(): string {
@@ -38,25 +32,6 @@ export class BookSearchComponent {
 
   addBookToReadingList(book: Book) {
     this.store.dispatch(addToReadingList({ book }));
-
-    this.store
-      .select(confirmedAddToReadingList)
-      .pipe(take(1))
-      .subscribe(() => {
-        const snackBarRef = this.snackbar.open(
-          book.title + BookConstant.ADDED_MSG,
-          BookConstant.UNDO,
-          {
-            duration: BookConstant.SNACKBAR_DURATION,
-          }
-        );
-
-        snackBarRef.onAction().subscribe(() => {
-          this.store.dispatch(
-            removeFromReadingList({ item: { ...book, bookId: book.id } })
-          );
-        });
-      });
   }
   searchExample() {
     this.searchForm.controls.term.setValue('javascript');
